requirements.txt
Django==5.0.2
python-dotenv==1.0.1
crispy-bootstrap5
django-crispy-forms 2.1

